refactor: データの入力に不備がある時の例外追加
