+++ src/main/java/com/example/moneyAllocation/exception/ResourceValidationException.java
@@ -1,0 +1,12 @@
+package com.example.moneyAllocation.exception;
+
+public class ResourceValidationException extends RuntimeException {
+    public ResourceValidationException(String message) {
+        super(message);
+    }
+
+    public ResourceValidationException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+}
